processor-service/requirements.txt
redis==5.0.1
pandas
scikit-learn
psycopg2-binary